feat(auth): add handleLogin controller for sign-in endpoint
