+++ server/src/controllers/auth/index.ts
@@ -1,1 +1,2 @@
 export * from "./handle-new-user";
+export * from "./handle-login";
